backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
# 2.0.36+ required for Python 3.13 (TypingOnly / __static_attributes__ fix)
sqlalchemy>=2.0.36
python-dotenv==1.0.0
openai==1.12.0
anthropic==0.18.1
# Use 2.10+ for Python 3.13 pre-built wheels (pydantic-core)
pydantic>=2.10.0,<3
pydantic-settings>=2.1.0
sqlparse==0.4.4
aiosqlite==0.19.0
httpx==0.26.0
bcrypt>=4.0.0
python-jose[cryptography]==3.3.0
